pyproject.toml
[tool.poetry]
name = "mis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pytest = "^8.3.4"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
rest-framework-simplejwt = "^0.0.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
